Moved ModManager.SubmitNewMod/ModChanges to a separate file
