+++ Runtime/ComplexCallbacks/ModManager_SubmitModOperation.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 421d4099868e24be2a1a8d6dbf584056
+timeCreated: 1594524787
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
